accomplish refactory routers
