+++ package.json
@@ -52,7 +52,7 @@
         "redis": "^3.0.2",
         "sinon": "^8.1.1",
         "socket.io": "^3.0.4",
-        "validator": "^13.0.0"
+        "validator": "^13.5.2"
     },
     "devDependencies": {
         "eslint": "^6.8.0"
